# openapi/baseline.json
{
  "components": {
    "schemas": {
...
      "CreateAnnotationRequest": {
        "properties": {
          "body": {
            "type": "string"
          },
          "snapshotId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "CustomPrivateRangeDto": {
        "properties": {
          "cidr": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "label": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "IpOrgRuleDto": {
        "properties": {
          "cidr": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "label": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "NetworkAnnotationDto": {
        "properties": {
          "body": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "networkId": {
            "format": "uuid",
            "type": "string"
          },
          "snapshotId": {
            "format": "uuid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
      "NetworkDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "criticalChanges": {
            "format": "int64",
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "hasInsights": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "snapshotCount": {
            "format": "int32",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "warningChanges": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "UpdateAnnotationRequest": {
        "properties": {
          "body": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "UpdateNetworkRequest": {
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/custom-private-ranges": {
      "get": {
        "operationId": "list_4",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CustomPrivateRangeDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List custom private ranges",
        "tags": [
          "Custom Private Ranges"
        ]
      },
      "post": {
        "operationId": "create_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomPrivateRangeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Add a custom private range",
        "tags": [
          "Custom Private Ranges"
        ]
      }
    },
    "/api/v1/custom-private-ranges/{id}": {
      "delete": {
        "operationId": "delete_6",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Delete a custom private range",
        "tags": [
          "Custom Private Ranges"
        ]
      }
    },
...
    "/api/v1/ip-org-rules": {
      "get": {
        "operationId": "list_3",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IpOrgRuleDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List IP organization rules",
        "tags": [
          "IP Org Rules"
        ]
      },
      "post": {
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IpOrgRuleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Add an IP organization rule",
        "tags": [
          "IP Org Rules"
        ]
      }
    },
    "/api/v1/ip-org-rules/{id}": {
      "delete": {
        "operationId": "delete_4",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Delete an IP organization rule",
        "tags": [
          "IP Org Rules"
        ]
      }
    },
...
    "/api/v1/monitor/networks/{networkId}": {
      "delete": {
        "operationId": "deleteNetwork",
        "parameters": [
          {
            "in": "path",
            "name": "networkId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "summary": "Delete a network",
        "tags": [
          "Monitor Networks"
        ]
      },
      "get": {
        "operationId": "getNetwork",
        "parameters": [
          {
            "in": "path",
            "name": "networkId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get a single network by ID",
        "tags": [
          "Monitor Networks"
        ]
      },
      "patch": {
        "operationId": "updateNetwork",
        "parameters": [
          {
            "in": "path",
            "name": "networkId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNetworkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Update a network's metadata",
        "tags": [
          "Monitor Networks"
        ]
      }
    },
    "/api/v1/monitor/networks/{networkId}/annotations": {
      "get": {
        "operationId": "list_2",
        "parameters": [
          {
            "in": "path",
            "name": "networkId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NetworkAnnotationDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List annotations for a network",
        "tags": [
          "Network Annotations"
        ]
      },
      "post": {
        "operationId": "create_1",
        "parameters": [
          {
            "in": "path",
            "name": "networkId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnnotationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkAnnotationDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Create an annotation",
        "tags": [
          "Network Annotations"
        ]
      }
    },
    "/api/v1/monitor/networks/{networkId}/annotations/{annotationId}": {
      "delete": {
        "operationId": "delete_3",
        "parameters": [
          {
            "in": "path",
            "name": "networkId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "annotationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Delete an annotation",
        "tags": [
          "Network Annotations"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "in": "path",
            "name": "networkId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "annotationId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAnnotationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkAnnotationDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Update an annotation's body",
        "tags": [
          "Network Annotations"
        ]
      }
    },
...
  },
...
}